Refactor some of the code and clean it up
- Also tested with a larger pivot table, seems to work well
