+++ plugins/ag-grid/src/js/package.json
@@ -49,15 +49,15 @@
     "react-dom": "^17.0.2 || ^18.0.0 || ^19.0.0"
   },
   "dependencies": {
-    "@ag-grid-community/core": "^32.3.4",
-    "@ag-grid-community/react": "^32.3.4",
-    "@ag-grid-community/styles": "^32.3.4",
-    "@ag-grid-community/theming": "^32.3.4",
-    "@ag-grid-enterprise/column-tool-panel": "^32.3.4",
-    "@ag-grid-enterprise/core": "^32.3.4",
-    "@ag-grid-enterprise/row-grouping": "^32.3.4",
-    "@ag-grid-enterprise/server-side-row-model": "^32.3.4",
-    "@ag-grid-enterprise/viewport-row-model": "^32.3.4",
+    "@ag-grid-community/core": "^32.3.9",
+    "@ag-grid-community/react": "^32.3.9",
+    "@ag-grid-community/styles": "^32.3.9",
+    "@ag-grid-community/theming": "^32.3.9",
+    "@ag-grid-enterprise/column-tool-panel": "^32.3.9",
+    "@ag-grid-enterprise/core": "^32.3.9",
+    "@ag-grid-enterprise/row-grouping": "^32.3.9",
+    "@ag-grid-enterprise/server-side-row-model": "^32.3.9",
+    "@ag-grid-enterprise/viewport-row-model": "^32.3.9",
     "@deephaven/components": "^0.106.2",
     "@deephaven/icons": "^0.106.0",
     "@deephaven/jsapi-bootstrap": "^0.106.2",
